requirements.txt
PyYAML
requests
scipy
